refactor: vo 패키지 제거
