+++ src/main/java/com/meoguri/linkocean/domain/linkmetadata/entity/LinkMetadata.java
@@ -7,7 +7,6 @@ import javax.persistence.Embedded;
 import javax.persistence.Entity;
 
 import com.meoguri.linkocean.domain.BaseIdEntity;
-import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
