+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/entity/LinkMetadataTest.java
@@ -5,8 +5,6 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Url;
-
 class LinkMetadataTest {
 
 	@Test
