+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/entity/UrlTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.linkmetadata.entity.vo;
+package com.meoguri.linkocean.domain.linkmetadata.entity;
 
 import static org.assertj.core.api.Assertions.*;
