+++ src/test/java/com/meoguri/linkocean/domain/user/entity/EmailTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.user.entity.vo;
+package com.meoguri.linkocean.domain.user.entity;
 
 import static org.assertj.core.api.Assertions.*;
