+++ src/test/java/com/meoguri/linkocean/domain/user/entity/UserTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
 import com.meoguri.linkocean.domain.user.entity.User.OAuthType;
-import com.meoguri.linkocean.domain.user.entity.vo.Email;
 
 class UserTest {
